% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_approach.R
\name{fit_bayes_reg_rfm}
\alias{fit_bayes_reg_rfm}
\title{Fit Bayesian regularization random features model}
\usage{
fit_bayes_reg_rfm(
  x,
  y,
  n_features = 100,
  weight_dist = "normal",
  weight_params = list(),
  bias_dist = "uniform",
  bias_params = list(min_val = 0, max_val = 2 * pi),
  act_func = "fourier",
  reg_type = "lasso",
  burnin = 1000,
  n.samples = 2000,
  thin = 5,
  n.cores = 1,
  pred_type = "mean",
  CI = 95
)
}
\arguments{
\item{x}{A matrix of features.}

\item{y}{A response vector.}

\item{n_features}{A positive integer of the number of random features. Defaults to \code{100}.}

\item{weight_dist}{A character string of the name of the distribution of the weight matrix.
Defaults to \code{"normal"}. Other supported options include \code{"uniform"}, \code{"cauchy"}, \code{"exponential"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{weight_params}{A list of the parameters for the distribution of the weight matrix.
\itemize{
\item If \code{weight_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{-1}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{weight_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to an empty list \code{list()} (which is interpreted differently based on \code{weight_dist}).}

\item{bias_dist}{A character string of the name of the distribution of the bias vector.
Defaults to \code{"uniform"}. Other options include \code{"normal"}, \code{"exponential"}, \code{"cauchy"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{bias_params}{A list of the parameters for the distribution of the bias vector.
\itemize{
\item If \code{bias_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{0}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{2*pi}.
}
\item If \code{bias_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{bias_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to a list of parameters of the uniform distribution \code{list(min_val=0, max_val=2*pi)}.}

\item{act_func}{A character string of the activation function of the random features model.
Possible values: \code{"fourier"} (default), \code{"sigmoid"}, \code{"tanh"}, \code{"sine"}, \code{"cosine"}, \code{"relu"}.}

\item{reg_type}{A character string of the regularization method.
Possible values: \code{"lasso"} (default), \code{"ridge"}.}

\item{burnin}{A numeric value specifying the number of burn-in samples. Defaults to \code{1000}.}

\item{n.samples}{A numeric value indicating the number of posterior samples to generate. Defaults to \code{2000}.}

\item{thin}{A numeric value suggesting the level of thinning. Defaults to \code{5}.}

\item{n.cores}{The number of cores for calculation defined by the \code{bayesreg} package. Defaults to \code{1}.}

\item{pred_type}{A character string indicating how to summarize the posterior predictive distribution.
Possible values: \code{"mean"} (default), \code{"median"}.}

\item{CI}{A numeric value indicating the level of the credible interval to report (in percentage). Defaults to \code{95}.}
}
\value{
A list of class \code{"BayesRandFeatRegModel"} containing the fitted model with components:

\code{fit_results}: The result of fitted model from \code{bayesreg::bayesreg}.

\code{W}: The matrix of random weights used of random features model.

\code{b}: The vector of random biases used of random features model.

\code{family}: A character string for the likelihood family.

\code{act_func}: A character string of the activation function used in random features model.

\code{CI}: The level of the credible interval to report.

\code{pred_type}: A character string of the prediction summary type (\code{"mean"} or \code{"median"}).

\code{train_preds}: A matrix containing the predicted values (\code{mean} or \code{median} depending on \code{pred_type}), along with the lower and upper bounds of the credible interval based on the specified \code{CI}.

\code{posterior_samples}: A matrix of posterior samples of model coefficients.

\code{posterior_sigma2}: A numeric vector of posterior samples of the error variance.

\code{ess}: An \code{"ess_plot"} object including effective sample sizes (ESS) for MCMC diagnostics.

\code{WAIC}: The Watanabe–Akaike information criterion of the fitted model.
}
\description{
Fit the regularization random features model under Bayesian framework.
Currently, only Gaussian distribution is supported.
The \code{bayesreg} package is used to estimate the model coefficients via penalized regression (lasso and ridge).
}
\details{
The model is defined as:
\deqn{\mathbf{y} = \beta_0+\sigma(\mathbf{XW} + \mathbf{b}) \boldsymbol{\beta}+\epsilon}
where \eqn{\sigma} is the activation function, \eqn{\mathbf{X}} is the input matrix, \eqn{\mathbf{W}} is the random weighting matrix, \eqn{\mathbf{b}} is the random bias vector,
\eqn{\beta_0} is the estimated intercept, \eqn{\boldsymbol{\beta}} is the learnt weights matrix using regularization techniques, and \eqn{\epsilon\sim \mathcal{N}(0, \sigma_{\epsilon}^2)} represents errors.

Note that \eqn{\beta_0}, \eqn{\boldsymbol{\beta}}, and \eqn{\sigma_{\epsilon}^2} are estimated using \code{bayesreg::bayesreg}.
}
\examples{
\dontrun{
set.seed(456)
x = cbind(rnorm(1000), runif(1000, min = -1, max = 1))
y = x[,1] + 2*x[,2] + x[,1]^2 + x[,1]^3

fitted <- fit_bayes_reg_rfm(x, y, n_features = 50)
}

}
\references{
\code{bayesreg}: Bayesian Regression Models with Global-Local Shrinkage Priors

CRAN package manual: \url{https://cran.r-project.org/web/packages/bayesreg/index.html}
}
