% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_approach.R
\name{fit_freq_reg_rfm}
\alias{fit_freq_reg_rfm}
\title{Fit frequentist regularization regression with random features}
\usage{
fit_freq_reg_rfm(
  x,
  y,
  n_features = 100,
  weight_dist = "normal",
  weight_params = list(),
  bias_dist = "uniform",
  bias_params = list(min_val = 0, max_val = 2 * pi),
  act_func = "fourier",
  standardize = TRUE,
  nfolds = 10,
  reg_type = "lasso",
  alpha = NULL,
  lambda_type = "min",
  lambda = c(10^-3, 10^-4, 10^-5, 10^-6, 10^-7)
)
}
\arguments{
\item{x}{A matrix of features.}

\item{y}{A response vector.}

\item{n_features}{An integer of the number of random features. Defaults to \code{100}.}

\item{weight_dist}{A character string of the name of the distribution of the weight matrix.
Defaults to \code{"normal"}. Other supported options include \code{"uniform"}, \code{"cauchy"}, \code{"exponential"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{weight_params}{A list of the parameters for the distribution of the weight matrix.
\itemize{
\item If \code{weight_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{-1}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{weight_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to an empty list \code{list()} (which is interpreted differently based on \code{weight_dist}).}

\item{bias_dist}{A character string of the name of the distribution of the bias vector.
Defaults to \code{"uniform"}. Other options include \code{"normal"}, \code{"exponential"}, \code{"cauchy"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{bias_params}{A list of the parameters for the distribution of the bias vector.
\itemize{
\item If \code{bias_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{0}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{2*pi}.
}
\item If \code{bias_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{bias_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to a list of parameters of the uniform distribution \code{list(min_val=0, max_val=2*pi)}.}

\item{act_func}{A character string of the activation function of the random features model.
Possible values: \code{"fourier"} (default), \code{"sigmoid"}, \code{"tanh"}, \code{"sine"}, \code{"cosine"}, \code{"relu"}.}

\item{standardize}{A logical flag to standardize the transformation of the random features model for \code{glmnet::glmnet()}. Defaults to \code{TRUE}.}

\item{nfolds}{An integer of the number of folds. Defaults to \code{10}.}

\item{reg_type}{A character string of the regularization method.
Possible values: \code{"lasso"} (default), \code{"ridge"}, \code{"elastic net"}.}

\item{alpha}{A numeric value for the elastic net mixing parameter.
Use \code{alpha = 1} for LASSO, \code{alpha = 0} for Ridge, \code{alpha = 0.5} (default) for Elastic net.
If \code{reg_type} is \code{"lasso"} or \code{"ridge"}, \code{alpha} can be set to \code{NULL}.
Defaults to \code{NULL}.}

\item{lambda_type}{A character string indicating the type of lambda to use after cross-validation.
Possible values are \code{"min"} (default), and \code{"1se"}.}

\item{lambda}{A numeric vector of possible lambda values to use in cross-validation. Defaults to \code{c(10^-3, 10^-4, 10^-5, 10^-6, 10^-7)}.}
}
\value{
A list of class \code{"FreqRandFeatRegModel"} containing the fitted model with components::

\code{fit_results}: The result of fitted model from \code{glmnet::glmnet}.

\code{lambda}: The optimal lambda value obtained from \code{glmnet::cv.glmnet}.

\code{W}: The matrix of random weights used of random features model.

\code{b}: The vector of random biases used of random features model.

\code{AIC}: The Akaike Information Criterion of the fitted model.

\code{cv_model}: The result of cross-validation model from \code{glmnet::cv.glmnet}.

\code{train_preds}: A numeric vector of predicted values for the training data.
}
\description{
The function aims to fit the regularization random features model using the frequentist approach.
Currently, only Gaussian distribution is supported.
The \code{glmnet} package is used to estimate the model coefficients via penalized regression.
}
\details{
The model is defined as:
\deqn{\mathbf{y} = \beta_0+\sigma(\mathbf{XW} + \mathbf{b}) \boldsymbol{\beta}+\epsilon}
where \eqn{\sigma} is the activation function, \eqn{\mathbf{X}} is the input matrix, \eqn{\mathbf{W}} is the random weighting matrix, \eqn{\mathbf{b}} is the random bias vector,
\eqn{\beta_0} is the estimated intercept, \eqn{\boldsymbol{\beta}} is the learnt weights matrix using regularization techniques, and \eqn{\epsilon\sim \mathcal{N}(0, \sigma_{\epsilon}^2)} represents errors.

Note that \eqn{\beta_0} and \eqn{\boldsymbol{\beta}} are estimated using \code{glmnet::glmnet}.
}
\examples{
\dontrun{
set.seed(456)
x = cbind(rnorm(1000), runif(1000, min = -1, max = 1))
y = x[,1] + 2*x[,2] + x[,1]^2 + x[,1]^3

fitted <- fit_freq_reg_rfm(x, y, n_features = 50)
}

}
\references{
\code{glmnet}: Lasso and Elastic-Net Regularized Generalized Linear Models

CRAN package manual: \url{https://cran.r-project.org/web/packages/glmnet/index.html}
}
