% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd_difference.R
\name{forward_difference}
\alias{forward_difference}
\title{Calculate Time-Series Forward Differences}
\usage{
forward_difference(x, step = 1)
}
\arguments{
\item{x}{A numeric vector.}

\item{step}{A numeric value or vector specifying time steps.
If a vector is provided, its length (after removing any NAs) must equal \code{length(x) - 1}.
Defaults to \code{1}.}
}
\description{
Calculates the forward differences (rate of change) for a time-series vector.
}
\details{
The rate of change is computed using the formula
\eqn{\frac{x_{t_2} - x_{t_1}}{t_2 - t_1}}, where \eqn{x_{t}} is the value at time \eqn{t}.
}
