% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_function.R
\name{plot.ess_plot}
\alias{plot.ess_plot}
\title{Plot method for Effective Sample Size Plot  objects}
\usage{
\method{plot}{ess_plot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"ess_plot"}.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Plot method for Effective Sample Size Plot  objects
}
