% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_approach.R
\name{predict.BayesRandFeatRegModel}
\alias{predict.BayesRandFeatRegModel}
\title{Predict function for \code{BayesRandFeatRegModel} objects}
\usage{
\method{predict}{BayesRandFeatRegModel}(object, newx, ...)
}
\arguments{
\item{object}{A \code{BayesRandFeatRegModel} object.}

\item{newx}{New input data (a matrix) for prediction.}

\item{...}{Additional arguments (ignored).}
}
\value{
A list containing:
\itemize{
\item \code{pred}: A numeric vector of predict values.
\item \code{lower}: A numeric vector of the lower bounds of the credible interval.
\item \code{upper}: A numeric vector of the upper bounds of the credible interval.
\item \code{full_samples}: A matrix of all posterior predictive samples.
}
}
\description{
The function aims to predict the response value given a new input newx.
}
