% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_approach.R
\name{predict.FreqRandFeatRegModel}
\alias{predict.FreqRandFeatRegModel}
\title{Predict function for FreqRandFeatRegModel objects}
\usage{
\method{predict}{FreqRandFeatRegModel}(object, newx, ...)
}
\arguments{
\item{object}{A FreqRandFeatRegModel object.}

\item{newx}{New input data (a matrix) for prediction.}

\item{...}{Additional arguments (ignored).}
}
\value{
A vector of predicted values.
}
\description{
The function aims to predict the response value given a new input newx using Random Features Model under the Frequentist Regularization approach.
}
