% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR_construction.R
\name{se_ts_freq_taken_reg_rfm}
\alias{se_ts_freq_taken_reg_rfm}
\title{Calculate standard error for confidence interval of time series data when fitting frequentist regularization regression with random features for time delay embedding using bootstrapping}
\usage{
se_ts_freq_taken_reg_rfm(
  ts_data,
  time,
  ref_point = NULL,
  smooth_diff = TRUE,
  method = "ma",
  smooth_params = list(),
  window_size = 9,
  pred_size = 7,
  pred_time_step = 1,
  weight_dist = "normal",
  weight_params = list(),
  bias_dist = "uniform",
  bias_params = list(min_val = 0, max_val = 2 * pi),
  act_func = "fourier",
  alpha = NULL,
  lambda = c(10^-3, 10^-4, 10^-5, 10^-6, 10^-7),
  lambda_type = "min",
  standardize = TRUE,
  nfolds = 10,
  reg_type = "lasso",
  feature_selection = "sqrt",
  feature_constant = NULL
)
}
\arguments{
\item{ts_data}{A numeric vector representing the time series data.}

\item{time}{A numeric vector representing the time gap between two consecutive observations in time series. Defaults to \code{NULL}.}

\item{ref_point}{An integer of the starting training point for calculating the standard error.
Defaults to \code{length(ts_data)/3}if it is greater than \code{pred_size}, otherwise it is \code{length(ts_data)-pred_size}.}

\item{smooth_diff}{A boolean value  to choose to smooth the time derivatives.
Possible values: \code{TRUE} (default), \code{T}, \code{FALSE}, \code{F}.}

\item{method}{A character string of the smooth method for the time derivatives.
Defaults to \code{ma}. Other supported options include \code{"polynomial"}, \code{"localized_polynomial"}, \code{"spline"}.}

\item{smooth_params}{A list of the parameters for the smooth function.
\itemize{
\item If \code{method = "ma"}, the list contains:
\itemize{
\item \code{window}: Window size to apply moving average. Defaults to \code{5}.
}
\item If \code{method = "polynomial"}, the list contains:
\itemize{
\item \code{polyorder}: The order of the polynomial. Defaults to \code{5}.
}
\item If \code{method = "localized_polynomial"}, the list contains:
\itemize{
\item \code{span}: The size how localized is the fit, which ranges between 0 to 1. Defaults to \code{0.5}.
}
\item If \code{method = "spline"}, the list contains:
\itemize{
\item \code{knotnum}: The number of knots. Defaults to \code{5}.
}
}

Defaults to an empty list \code{list()} (which is interpreted differently based on \code{method}).}

\item{window_size}{An integer of the embedding dimension of time series. Defaults to \code{9}.}

\item{pred_size}{An integer of the prediction time steps. Defaults to \code{7}.}

\item{pred_time_step}{An integer or a vector of time gap between two consecutive observations. Defaults to \code{1}.}

\item{weight_dist}{A character string of the name of the distribution of the weight matrix.
Defaults to \code{"normal"}. Other supported options include \code{"uniform"}, \code{"cauchy"}, \code{"exponential"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{weight_params}{A list of the parameters for the distribution of the weight matrix.
\itemize{
\item If \code{weight_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{-1}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{weight_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to an empty list \code{list()} (which is interpreted differently based on \code{weight_dist}).}

\item{bias_dist}{A character string of the name of the distribution of the bias vector.
Defaults to \code{"uniform"}. Other options include \code{"normal"}, \code{"exponential"}, \code{"cauchy"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{bias_params}{A list of the parameters for the distribution of the bias vector.
\itemize{
\item If \code{bias_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{0}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{2*pi}.
}
\item If \code{bias_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{bias_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to a list of parameters of the uniform distribution \code{list(min_val=0, max_val=2*pi)}.}

\item{act_func}{A character string of the activation function of the random features model.
Possible values: \code{"fourier"} (default), \code{"sigmoid"}, \code{"tanh"}, \code{"sine"}, \code{"cosine"}, \code{"relu"}.}

\item{alpha}{A numeric value for the elastic net mixing parameter.
Use \code{alpha = 1} for LASSO, \code{alpha = 0} for Ridge, \code{alpha = 0.5} (default) for Elastic net.
If \code{reg_type} is \code{"lasso"} or \code{"ridge"}, \code{alpha} can be set to \code{NULL}.
Defaults to \code{NULL}.}

\item{lambda}{A numeric vector of possible lambda values to use in cross-validation. Defaults to \code{c(10^-3, 10^-4, 10^-5, 10^-6, 10^-7)}.}

\item{lambda_type}{A character string indicating the type of lambda to use after cross-validation.
Possible values are \code{"min"} (default), and \code{"1se"}.}

\item{standardize}{A logical flag to standardize the transformation of the random features model. Defaults to \code{TRUE}.}

\item{nfolds}{An integer of the number of folds. Defaults to \code{10}.}

\item{reg_type}{A character string of the regularization method.
Possible values: \code{"lasso"} (default), \code{"ridge"}, \code{"elastic net"}.}

\item{feature_selection}{A character string specifying the method for selecting the number of random features.
Possible values are:
\itemize{
\item \code{"sqrt"} (default): The number of features is set to the square root of the number of rows of the training matrix \code{nrow}, i.e., \code{sqrt(nrow)}.
\item \code{"factor"}: The number of features is \code{feature_constant*nrow}.
\item \code{"constant"}: The number of features is fixed at \code{feature_constant}.
}}

\item{feature_constant}{A numeric value used when \code{feature_selection} is \code{"factor"} or \code{"constant"}.
Ignored when \code{feature_selection = "sqrt"}. Defaults to \code{NULL}.}
}
\value{
A list containing standard error and relevant components:
\itemize{
\item \code{se}: The standard error of the predictions.
\item \code{pred_mat}: A matrix containing predictions, where the columns represent the prediction size.
\item \code{mat_true}: A matrix containing true values of the input time series, where the columns represent the prediction size.
}
}
\description{
This function aims to compute the standard error of model prediction given the time series data using bootstrapping.
}
\examples{
\dontrun{
set.seed(456)
y = cumsum(rnorm(1000))
res_se <- se_ts_freq_taken_reg_rfm(ts_data = y, ref_point = 900)
}

}
