% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness_function.R
\name{smoothness_ts}
\alias{smoothness_ts}
\title{Smooth Time Series Data}
\usage{
smoothness_ts(ts_data, fn_name = "ma", params = list())
}
\arguments{
\item{ts_data}{A numeric vector representing the time series data to be smoothed.}

\item{fn_name}{A character string indicating the smoothing method to be used. Options include:
"polynomial", "ma", "localized_polynomial", "spline". Default is "ma".}

\item{params}{A list of additional parameters for the selected smoothing method.
E.g., for "polynomial", use \code{params$polyorder}; for "ma", use \code{params$window};
for "localized_polynomial" use \code{params$span}; and for "spline" use \code{params$knotnum}.}
}
\value{
A numeric vector of the smoothed time series data.
}
\description{
The function applies a smoothing technique to a time series using different methods including
polynomial smoothing, low-pass filtering, localized polynomial smoothing, or spline smoothing.
}
\examples{
# Example with moving average
\dontrun{
y = cumsum(rnorm(1000))
smoothed_data <- smoothness_ts(
  ts_data = y,
  fn_name = "ma",
  params = list(window = 5))
}
}
