% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_approach.R
\name{ts_forecast_bayes_reg_rfm}
\alias{ts_forecast_bayes_reg_rfm}
\title{Forecasting for time series using Bayesian regularization regression with random features model}
\usage{
ts_forecast_bayes_reg_rfm(
  ts_data,
  window_size = 9,
  pred_size = 7,
  weight_dist = "normal",
  weight_params = list(),
  bias_dist = "uniform",
  bias_params = list(min_val = 0, max_val = 2 * pi),
  act_func = "fourier",
  reg_type = "lasso",
  burnin = 1000,
  n.samples = 2000,
  thin = 5,
  n.cores = 1,
  pred_type = "mean",
  feature_selection = "sqrt",
  feature_constant = NULL,
  CI = 95
)
}
\arguments{
\item{ts_data}{A numeric vector representing the time series data.}

\item{window_size}{An integer specifying the embedding window size (i.e., the number of past observations used as features). Defaults to \code{9}.}

\item{pred_size}{An integer specifying the number of future time steps to predict. Defaults to \code{7}.}

\item{weight_dist}{A character string of the name of the distribution of the weight matrix.
Defaults to \code{"normal"}. Other supported options include \code{"uniform"}, \code{"cauchy"}, \code{"exponential"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{weight_params}{A list of the parameters for the distribution of the weight matrix.
\itemize{
\item If \code{weight_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{-1}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{weight_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to an empty list \code{list()} (which is interpreted differently based on \code{weight_dist}).}

\item{bias_dist}{A character string of the name of the distribution of the bias vector.
Defaults to \code{"uniform"}. Other options include \code{"normal"}, \code{"exponential"}, \code{"cauchy"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{bias_params}{A list of the parameters for the distribution of the bias vector.
\itemize{
\item If \code{bias_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{0}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{2*pi}.
}
\item If \code{bias_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{bias_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to a list of parameters of the uniform distribution \code{list(min_val=0, max_val=2*pi)}.}

\item{act_func}{A character string of the activation function of the random features model.
Possible values: \code{"fourier"} (default), \code{"sigmoid"}, \code{"tanh"}, \code{"sine"}, \code{"cosine"}, \code{"relu"}.}

\item{reg_type}{A character string of the regularization method.
Possible values: \code{"lasso"} (default), \code{"ridge"}.}

\item{burnin}{A numeric value specifying the number of burn-in samples. Defaults to \code{1000}.}

\item{n.samples}{A numeric value indicating the number of posterior samples to generate. Defaults to \code{2000}.}

\item{thin}{A numeric value suggesting the level of thinning. Defaults to \code{5}.}

\item{n.cores}{The number of cores for calculation defined by the \code{bayesreg} package. Defaults to \code{1}.}

\item{pred_type}{A character string indicating how to summarize the posterior predictive distribution.
Possible values: \code{"mean"} (default), \code{"median"}.}

\item{feature_selection}{A character string specifying the method for selecting the number of random features.
Possible values are:
\itemize{
\item \code{"sqrt"} (default): The number of features is set to the square root of the number of rows of the training matrix \code{nrow}, i.e., \code{sqrt(nrow)}.
\item \code{"factor"}: The number of features is \code{feature_constant*nrow}.
\item \code{"constant"}: The number of features is fixed at \code{feature_constant}.
Note that the number of features will be always rounded down.
}}

\item{feature_constant}{A numeric value used when \code{feature_selection} is \code{"factor"} or \code{"constant"}.
Ignored when \code{feature_selection = "sqrt"}. Defaults to \code{NULL}.}

\item{CI}{A numeric value indicating the level of the credible interval to report (in percentage). Defaults to \code{95}.}
}
\value{
A list containing the fitted model and relevant components:

\code{fit_results}: The result of fitted model from \code{bayesreg::bayesreg}.

\code{fitted.values}: The fitted mean or median values defined by \code{pred_type}.

\code{W}: The matrix of random weights used of random features model.

\code{b}: The vector of random biases used of random features model.

\code{y_pred}: A numeric vector of forecast values for future time steps, with length equal to \code{pred_size}.

\code{future_preds}: A matrix of all posterior predictive samples for future time steps.

\code{n_features}: A numeric value of the number of random features.

\code{posterior_samples}: A matrix of posterior samples of model coefficients.

\code{posterior_sigma2}: A numeric vector of posterior samples of the error variance.

\code{ess}: An \code{"ess_plot"} object including effective sample sizes (ESS) for MCMC diagnostics.

\code{WAIC}: The Watanabe–Akaike information criterion of the fitted model.

\code{pred.ci}: A matrix including the credible interval based on the input of \code{CI}.
}
\description{
The Bayesian regularization regression with random feature has been applied to predict the values of the time series with the specifying value of time steps.
Currently, only Gaussian distribution is supported.
The \code{bayesreg} package is used to estimate the model coefficients via penalized regression.
}
\details{
Given a time series \eqn{\{(y_{t_k})\}_{k=1}^n}, the embedding matrix with the window size of \eqn{m} is constructed as follows:
\deqn{
\mathbf{X} \gets
\begin{bmatrix}
y_{t_1} & y_{t_2} & \dots & y_{t_{m-1}} & y_{t_m} \\
y_{t_2} & y_{t_3} & \dots & y_{t_m} & y_{t_{m+1}} \\
\vdots & \vdots & & \vdots & \vdots \\
y_{t_{n-m}} & y_{t_{n-m+1}} & \dots & y_{t_{n-2}} & y_{t_{n-1}}
\end{bmatrix}
}
The corresponding output vector is defined as:
\deqn{
\mathbf{y} \gets
\begin{bmatrix}
y_{t_{m+1}} \\
y_{t_{m+2}} \\
\vdots \\
y_{t_{n}}
\end{bmatrix}
}
The model is defined as:
\deqn{\mathbf{y} = \beta_0+\sigma(\mathbf{XW} + \mathbf{b}) \boldsymbol{\beta}+\epsilon}
where \eqn{\sigma} is the activation function, \eqn{\mathbf{X}} is the input matrix, \eqn{\mathbf{W}} is the random weighting matrix, \eqn{\mathbf{b}} is the random bias vector,
\eqn{\beta_0} is the estimated intercept, \eqn{\boldsymbol{\beta}} is the learnt weights matrix using regularization techniques, and \eqn{\epsilon\sim \mathcal{N}(0, \sigma_{\epsilon}^2)} represents errors.

Note that \eqn{\beta_0}, \eqn{\boldsymbol{\beta}}, and \eqn{\sigma_{\epsilon}^2} are estimated using \code{bayesreg::bayesreg}.
}
\examples{
\dontrun{
set.seed(456)
y = cumsum(rnorm(1000))
fitted_bayes_reg_rfm <- ts_forecast_bayes_reg_rfm(ts_data=y)
}

}
\references{
\code{bayesreg}: Bayesian Regression Models with Global-Local Shrinkage Priors

CRAN package manual: \url{https://cran.r-project.org/web/packages/bayesreg/index.html}
}
