% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_taken_ts.R
\name{ts_forecast_freq_reg_rfm_taken}
\alias{ts_forecast_freq_reg_rfm_taken}
\title{Forecast time series data using Takens' theorem under the frequentist regularization regression with random features model}
\usage{
ts_forecast_freq_reg_rfm_taken(
  ts_data,
  time = NULL,
  smooth_diff = TRUE,
  method = "ma",
  smooth_params = list(),
  window_size = 9,
  pred_size = 7,
  pred_time_step = 1,
  weight_dist = "normal",
  weight_params = list(),
  bias_dist = "uniform",
  bias_params = list(min_val = 0, max_val = 2 * pi),
  act_func = "fourier",
  alpha = NULL,
  lambda = c(10^-3, 10^-4, 10^-5, 10^-6, 10^-7),
  lambda_type = "min",
  standardize = TRUE,
  nfolds = 10,
  reg_type = "lasso",
  feature_selection = "sqrt",
  feature_constant = NULL,
  bootstrap_CI = FALSE,
  bootstrap_ref_point = NULL,
  CI = 95
)
}
\arguments{
\item{ts_data}{A numeric vector representing the time series data.}

\item{time}{A numeric vector representing the time gap between two consecutive observations in time series. Defaults to \code{NULL}.}

\item{smooth_diff}{A boolean value for smoothing the time derivatives.
Possible values: \code{TRUE} (default), \code{T}, \code{FALSE}, \code{F}.}

\item{method}{A character string of the smooth method for the time derivatives.
Defaults to \code{ma} for the right-aligned moving average lags. Other supported options include \code{"polynomial"}, \code{"localized_polynomial"}, \code{"spline"}.}

\item{smooth_params}{A list of the parameters for the smooth function.
\itemize{
\item If \code{method = "ma"}, the list contains:
\itemize{
\item \code{window}: Window size to apply moving average. Defaults to \code{5}
}
\item If \code{method = "polynomial"}, the list contains:
\itemize{
\item \code{polyorder}: The order of the polynomial. Defaults to \code{5}
}
\item If \code{method = "localized_polynomial"}, the list contains:
\itemize{
\item \code{span}: The size how localized is the fit, which ranges between 0 to 1. Defaults to \code{0.5}.
}
\item If \code{method = "spline"}, the list contains:
\itemize{
\item \code{knotnum}: The number of knots. Defaults to \code{5}.
}
}

Defaults to an empty list \code{list()} (which is interpreted differently based on \code{method}).}

\item{window_size}{An integer specifying the embedding window size (i.e., the number of past observations used as features). Defaults to \code{9}.}

\item{pred_size}{An integer specifying the number of future time steps to predict. Defaults to \code{7}.}

\item{pred_time_step}{An integer or a vector of time gap between two consecutive observations.
The vector must have length the same as \code{length(rate_of_change)}. Defaults to \code{1}.}

\item{weight_dist}{A character string of the name of the distribution of the weight matrix.
Defaults to \code{"normal"}. Other supported options include \code{"uniform"}, \code{"cauchy"}, \code{"exponential"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{weight_params}{A list of the parameters for the distribution of the weight matrix.
\itemize{
\item If \code{weight_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{-1}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{weight_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{weight_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to an empty list \code{list()} (which is interpreted differently based on \code{weight_dist}).}

\item{bias_dist}{A character string of the name of the distribution of the bias vector.
Defaults to \code{"uniform"}. Other options include \code{"normal"}, \code{"exponential"}, \code{"cauchy"}, \code{"bernoulli"}, \code{"lognormal"}.}

\item{bias_params}{A list of the parameters for the distribution of the bias vector.
\itemize{
\item If \code{bias_dist = "normal"}, the list contains:
\itemize{
\item \code{mean}: Mean of the normal distribution (numeric). Defaults to \code{0}.
\item \code{sd}: Standard deviation of the normal distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "uniform"}, the list contains:
\itemize{
\item \code{min_val}: Minimum value for the uniform distribution (numeric). Defaults to \code{0}.
\item \code{max_val}: Maximum value for the uniform distribution (numeric). Defaults to \code{2*pi}.
}
\item If \code{bias_dist = "cauchy"}, the list contains:
\itemize{
\item \code{location}: Location parameter of the Cauchy distribution (numeric). Defaults to \code{0}.
\item \code{scale}: Scale parameter of the Cauchy distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "exponential"}, the list contains:
\itemize{
\item \code{rate}: Rate parameter of the exponential distribution (numeric). Defaults to \code{1}.
}
\item If \code{bias_dist = "bernoulli"}, the list contains:
\itemize{
\item \code{prob}: Probability value of the Bernoulli distribution (numeric). Defaults to \code{0.5}.
}
\item If \code{bias_dist = "lognormal"}, the list contains:
\itemize{
\item \code{meanlog}: Mean of the distribution of the log scale (numeric). Defaults to \code{0}.
\item \code{sdlog}: Standard deviation of the distribution of the log scale (numeric). Defaults to \code{1}.
}
}

Defaults to a list of parameters of the uniform distribution \code{list(min_val=0, max_val=2*pi)}.}

\item{act_func}{A character string of the activation function of the random features model.
Possible values: \code{"fourier"} (default), \code{"sigmoid"}, \code{"tanh"}, \code{"sine"}, \code{"cosine"}, \code{"relu"}.}

\item{alpha}{A numeric value for the elastic net mixing parameter.
Use \code{alpha = 1} for LASSO, \code{alpha = 0} for Ridge, \code{alpha = 0.5} (default) for Elastic net.
If \code{reg_type} is \code{"lasso"} or \code{"ridge"}, \code{alpha} can be set to \code{NULL}.
Defaults to \code{NULL}.}

\item{lambda}{A numeric vector of possible lambda values to use in cross-validation. Defaults to \code{c(10^-3, 10^-4, 10^-5, 10^-6, 10^-7)}.}

\item{lambda_type}{A character string indicating the type of lambda to use after cross-validation.
Possible values are \code{"min"} (default), and \code{"1se"}.}

\item{standardize}{A logical flag to standardize the transformation of the random feature model. Defaults to \code{TRUE}.}

\item{nfolds}{An integer of the number folds. Defaults to \code{10}.}

\item{reg_type}{A character string of the regularization method.
Possible values: \code{"lasso"} (default), \code{"ridge"}, \code{"elastic net"}.}

\item{feature_selection}{A character string specifying the method for selecting the number of random features.
Possible values are:
\itemize{
\item \code{"sqrt"} (default): The number of features is set to the square root of the number of rows of the training matrix \code{nrow}, i.e., \code{sqrt(nrow)}.
\item \code{"factor"}: The number of features is \code{feature_constant*nrow}.
\item \code{"constant"}: The number of features is fixed at \code{feature_constant}.
Note that the number of features will be always rounded down.
}}

\item{feature_constant}{A numeric value used when \code{feature_selection} is \code{"factor"} or \code{"constant"}.
Ignored when \code{feature_selection = "sqrt"}. Defaults to \code{NULL}.}

\item{bootstrap_CI}{A boolean value for constructing confidence intervals using bootstrapping. Defaults to \code{FALSE}.}

\item{bootstrap_ref_point}{A reference point in the training time series for bootstrapping Defaults to \code{NULL}.
If \code{bootstrap_CI = TRUE}. \code{bootstrap_ref_point = length(ts_data)/3} (default) if it is greater than \code{pred_size}, otherwise it is \code{length(ts_data)-pred_size}.}

\item{CI}{A numeric value indicating the level of the confidence interval to report (in percentage). Defaults to \code{95}.}
}
\value{
A list of predicted values generated using Takens' theorem, based on the frequentist regularization regression with random features model,
specifying for time series data with the smooothness rate of change:

\code{fit_results}: The result of fitted model from \code{glmnet::glmnet}.

\code{W}: The matrix of random weights used of random features model.

\code{b}: The vector of random biases used of random features model.

\code{AIC}: The Akaike Information Criterion of the fitted model.

\code{cv_model}: The result of cross-validation model from \code{cv.glmnet}.

\code{y_pred}: A numeric vector of forecast values for future time steps, with length equal to \code{pred_size}.

\code{pred.ci}: A matrix including the confidence interval using bootstrapping.

\code{n_features}: A numeric value of the number of random features.

\code{raw_diff}: A vector of the time derivatives.

\code{smooth_diff}: A vector of the smooth time derivatives.
}
\description{
This function aims to forecast time series data for a specified number of time steps (\code{pred_size}) using frequentist regularization regression with random features model.
Currently, only Gaussian distribution is supported.
The \code{glmnet} package is used to estimate the model coefficients via penalized regression.
}
\details{
Given a time series \eqn{\{(y_{t_k})\}_{k=1}^n}, the embedding matrix with the window size of \eqn{m} is constructed as follows:
\deqn{
\mathbf{X} \gets
\begin{bmatrix}
y_{t_1} & y_{t_2} & \dots & y_{t_{m-1}} & y_{t_m} \\
y_{t_2} & y_{t_3} & \dots & y_{t_m} & y_{t_{m+1}} \\
\vdots & \vdots & & \vdots & \vdots \\
y_{t_{n-m}} & y_{t_{n-m+1}} & \dots & y_{t_{n-2}} & y_{t_{n-1}}
\end{bmatrix}
}
The corresponding output vector is defined as:
\deqn{
\bar{\mathbf{y}}'\gets
\begin{bmatrix}
   \bar{y}'_{t_m}\\
   \bar{y}'_{t_{m+1}}\\
   \vdots\\
   \bar{y}'_{t_{n-1}}
\end{bmatrix}
}
where \eqn{\bar{y}'_{t_k}} is calculated using smooth function defined by \code{smooth_params} when \code{smooth_diff = c(TRUE, T)}
with the input of \deqn{y'_{t_k} \approx \frac{y_{t_{k+1}}-y_{t_k}}{t_{k+1}-t_k}} for \eqn{k=1, \dots, n-1}.

The model is defined as:
\deqn{\bar{\mathbf{y}}' = \beta_0+\sigma(\mathbf{XW} + \mathbf{b}) \boldsymbol{\beta}+\epsilon}
where \eqn{\sigma} is the activation function, \eqn{\mathbf{X}} is the input matrix as defined above, \eqn{\mathbf{W}} is the random weighting matrix, \eqn{\mathbf{b}} is the random bias vector,
\eqn{\beta_0} is the estimated intercept, \eqn{\boldsymbol{\beta}} is the learnt weights matrix using regularization techniques, and \eqn{\epsilon\sim \mathcal{N}(0, \sigma_{\epsilon}^2)} represents errors.

Note that \eqn{\beta_0} and \eqn{\boldsymbol{\beta}} are estimated using \code{glmnet::glmnet}.

The forecast value of the time series is computed as:
\deqn{y_{t_{k+1}}\gets y_{t_k}+y'_{t_k}(t_{k+1}-t_k)}
for \eqn{k=1, \dots, n-1}.
}
\examples{
\dontrun{
set.seed(456)
y = cumsum(rnorm(1000))
fitted_ts_freq_taken_reg_rfm <- ts_forecast_freq_reg_rfm_taken(
    ts_data=y, bootstrap_CI=TRUE,
    bootstrap_ref_point=900)
}

}
\references{
\code{glmnet}: Lasso and Elastic-Net Regularized Generalized Linear Models

CRAN package manual: \url{https://cran.r-project.org/web/packages/glmnet/index.html}
}
