% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CR_construction.R
\name{upper_lower_ci}
\alias{upper_lower_ci}
\title{Confidence Interval Construction for Time series predictions}
\usage{
upper_lower_ci(pred, se, pred_size, critical_value = qnorm(0.975))
}
\arguments{
\item{pred}{A numeric vector of the predictions.}

\item{se}{A numeric vector of the standard errors corresponding to the prediction time steps.}

\item{pred_size}{A positive integer of the prediction time steps.}

\item{critical_value}{A numeric value of the critical value for constructing confidence intervals. Defaults to \code{qnorm(0.975)}}
}
\value{
A list contains the lower and upper bounds of the confidence intervals:

\code{lower}: The lower bound of the confidence intervals.

\code{upper}: The upper bound of the confidence intervals.
}
\description{
This function helps to produce the confidence interval by inputting the predicted values, the standard errors, the prediction time steps and the critical value.
}
